Moved several API functions from the authContext into firebase-api.js
